requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-speech
google-cloud-storage